seperate out components from FullBar
