+++ src/BarContentContainerTest.tsx
@@ -88,11 +88,14 @@ const App = () => {
                 </NumberInput>
                 <div id={"Bar-and-dec-test"} style={{width: "100%", height: "100%"}}>
                     <div id={"full_bar_plot-1"} style={{width: "100%", height: "200px"}}>
-                        <BarContentContainer
+
+                        {/* Props changed to work with legend For loop component */}
+
+                        {/* <BarContentContainer
                             elements={contentElements}
                             CSS="padding-right: 1rem;" 
                             decorationWidth="100%" 
-                        />
+                        /> */}
                     </div>
                     {`Data: ` + data.get()?.length??`None`}
                     {`\nDataMax: ` + dataMax.get()??`None`}
